Complete 2022 Data Conversion .mat - .h5
A compilation of data conversion script from .mat to .h5 for all data taken in 2022.  Dates include 01/12, 02/23, 03/17, 06/20, 07/15, 07/20 and 09/23.  Renaming this script to "Full BMX Catalogue" and adding future data conversion scripts will allow a running log of future data collected to be added to the dictionary file.

Time-averaged velocity data is included for dates including and after 3/17 within each dataset.